requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
asn1crypto==1.2.0
astropy==4.0.1.post1
astropy-healpix==0.5
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cachetools==4.1.0
camb==1.3.0
certifi==2020.6.20
cffi==1.13.0
chardet==3.0.4
classy==2.9.4
cloudpickle==1.4.1
conda==4.8.4
conda-package-handling==1.6.0
corner==2.1.0
cosmoplotian==0.0.1
cryptography==2.8
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
dm-sonnet==2.0.0
dm-tree==0.1.5
dustvaeder==0.1.0
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
GetDist==1.1.2
gin-config==0.4.0
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
graphviz==0.14.1
grpcio==1.29.0
gviz-api==1.9.0
h5py==2.10.0
healpy==1.14.0
horovod==0.19.0
idna==2.8
importlib-metadata==1.6.0
importlib-resources==5.1.0
ipykernel==5.1.4
ipyparallel==6.2.4
ipython==7.13.0
ipython-genutils==0.2.0
jax==0.2.9
jax-cosmo==0.1rc7
jaxlib==0.1.60
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyter-server-proxy==1.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
ligo-gracedb==2.7.4
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mpi4py==3.0.0
mpmath==1.2.1
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.6.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-nersc-customs==0.2.0
python3-nersc-modster==0.3.0
python3-nersc-sitecustomize==0.2.0
python3-nerscjson==0.9
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
recfast4py==0.2.2
reproject==0.7.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.46
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
setuptools-scm==5.0.1
simpervisor==0.3
six==1.12.0
sympy==1.7.1
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-profile==2.3.0rc0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-metadata==0.27.0
tensorflow-probability==0.10.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.36.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.24.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.16.0
yarl==1.4.2
zipp==3.1.0
